Move sync actions to appSlice
