+++ packages/desktop-client/src/components/App.tsx
@@ -17,9 +17,8 @@ import {
   loadBudget,
   loadGlobalPrefs,
   signOut,
-  sync,
 } from 'loot-core/client/actions';
-import { setAppState } from 'loot-core/client/app/appSlice';
+import { setAppState, sync } from 'loot-core/client/app/appSlice';
 import { SpreadsheetProvider } from 'loot-core/client/SpreadsheetProvider';
 import * as Platform from 'loot-core/src/client/platform';
 import {
