+++ packages/desktop-client/src/components/settings/Reset.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Trans } from 'react-i18next';
 
-import { resetSync } from 'loot-core/client/actions';
+import { resetSync } from 'loot-core/client/app/appSlice';
 import { send } from 'loot-core/src/platform/client/fetch';
 
 import { useMetadataPref } from '../../hooks/useMetadataPref';
