+++ packages/loot-core/src/client/actions/index.ts
@@ -3,5 +3,4 @@ export * from './notifications';
 export * from './prefs';
 export * from './budgets';
 export * from './backups';
-export * from './sync';
 export * from './user';
